+++ laokou-common/laokou-common-grpc-server/checkstyle/checkstyle-header.txt
@@ -1,0 +1,18 @@
+^\Q/*\E$
+^\Q * Copyright (c) \E2022-2026\Q KCloud-Platform-IoT Author or Authors. All Rights Reserved.\E$
+^\Q *\E$
+^\Q * Licensed under the Apache License, Version 2.0 (the "License");\E$
+^\Q * you may not use this file except in compliance with the License.\E$
+^\Q * You may obtain a copy of the License at\E$
+^\Q *\E$
+^\Q *   http://www.apache.org/licenses/LICENSE-2.0\E$
+^\Q *\E$
+^\Q * Unless required by applicable law or agreed to in writing, software\E$
+^\Q * distributed under the License is distributed on an "AS IS" BASIS,\E$
+^\Q * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.\E$
+^\Q * See the License for the specific language governing permissions and\E$
+^\Q * limitations under the License.\E$
+^\Q *\E$
+^\Q */\E$
+^$
+^.*$
