+++ laokou-common/laokou-common-grpc-server/checkstyle/checkstyle-suppressions.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0"?>
+<!--
+/*
+ * Copyright (c) 2022-2026 KCloud-Platform-IoT Author or Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+-->
+<!DOCTYPE suppressions PUBLIC
+  "-//Checkstyle//DTD SuppressionFilter Configuration 1.2//EN"
+  "https://checkstyle.org/dtds/suppressions_1_2.dtd">
+<suppressions>
+  <!-- https://checkstyle.sourceforge.io/filters/suppressionfilter.html -->
+  <suppress files="[/\\]target[/\\]" checks=".*"/>
+  <suppress files="[\\/]org[\\/]laokou[\\/]oss[\\/]api[\\/]" checks=".*" />
+  <suppress files="[\\/]org[\\/]laokou[\\/]distributed[\\/]identifier[\\/]api[\\/]" checks=".*" />
+</suppressions>
