+++ laokou-common/laokou-common-grpc-server/checkstyle/checkstyle.xml
@@ -1,0 +1,92 @@
+<?xml version="1.0"?>
+<!--
+/*
+ * Copyright (c) 2022-2026 KCloud-Platform-IoT Author or Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+-->
+<!DOCTYPE module PUBLIC
+		"-//Puppy Crawl//DTD Check Configuration 1.3//EN"
+		"https://www.puppycrawl.com/dtds/configuration_1_3.dtd">
+<module name="com.puppycrawl.tools.checkstyle.Checker">
+
+  <!-- https://checkstyle.sourceforge.io/property_types.html -->
+
+  <!-- 字符集 -->
+	<property name="charset" value="UTF-8"/>
+
+	<!-- 检查级别 -->
+	<property name="severity" value="error"/>
+
+<!--
+  <module name="io.spring.javaformat.checkstyle.SpringChecks">
+    <property name="excludes" value="io.spring.javaformat.checkstyle.check.SpringAvoidStaticImportCheck" />
+  </module>
+-->
+
+  <!-- 抑制过滤器 https://checkstyle.sourceforge.io/filters/suppressionfilter.html#SuppressionFilter -->
+  <module name="SuppressionFilter">
+    <property name="file" value="checkstyle/checkstyle-suppressions.xml"/>
+    <property name="optional" value="false"/>
+  </module>
+
+  <!-- 检查协议头 https://checkstyle.sourceforge.io/checks/header/regexpheader.html -->
+  <module name="RegexpHeader">
+    <property name="headerFile" value="checkstyle/checkstyle-header.txt"/>
+    <property name="fileExtensions" value="java"/>
+  </module>
+
+  <module name="TreeWalker">
+
+    <!-- 匹配句子末尾的格式 https://checkstyle.sourceforge.io/checks/javadoc/javadocstyle.html#JavadocStyle -->
+    <module name="JavadocStyleCheck">
+      <property name="endOfSentenceFormat" value="([.?!。？！][ \t\n\r\f&lt;])|([.?!。？！]$)"/>
+      <property name="checkHtml" value="false"/>
+    </module>
+
+    <!-- 禁止导入未使用的依赖 https://checkstyle.sourceforge.io/checks/imports/unusedimports.html -->
+    <module name="UnusedImports">
+      <!-- true忽略文档依赖，false不忽略文档依赖 -->
+      <property name="processJavadoc" value="true"/>
+    </module>
+
+    <!-- 禁止使用System.out.println https://checkstyle.sourceforge.io/checks/regexp/regexp.html -->
+    <module name="Regexp">
+      <property name="format" value="System\.out\.println"/>
+      <property name="illegalPattern" value="true"/>
+    </module>
+
+    <!-- 禁止使用System.out.print https://checkstyle.sourceforge.io/checks/regexp/regexp.html -->
+    <module name="Regexp">
+      <property name="format" value="System\.out\.print"/>
+      <property name="illegalPattern" value="true"/>
+    </module>
+
+    <!-- 禁止非法导入 https://checkstyle.sourceforge.io/checks/imports/illegalimport.html -->
+    <module name="IllegalImport">
+      <property name="illegalClasses" value="lombok.SneakyThrows"/>
+      <property name="illegalClasses" value="org.springframework.beans.factory.annotation.Autowired"/>
+      <property name="illegalClasses" value="org.junit.jupiter.api.Assertions.*"/>
+    </module>
+
+    <!-- 禁止静态导入 https://checkstyle.sourceforge.io/checks/imports/avoidstaticimport.html -->
+    <module name="AvoidStaticImport"/>
+
+    <!-- 禁止导入* https://checkstyle.sourceforge.io/checks/imports/avoidstarimport.html -->
+    <module name="AvoidStarImport"/>
+
+  </module>
+
+</module>
