+++ laokou-common/laokou-common-grpc-server/spotbugs/spotbugs-exclude.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<FindBugsFilter>
+
+  <!-- 排除某个特定 bug 模式的所有实例 -->
+  <Match>
+    <Bug pattern="BC_VACUOUS_INSTANCEOF,UWF_UNWRITTEN_PUBLIC_OR_PROTECTED_FIELD,SE_TRANSIENT_FIELD_NOT_RESTORED,NP_NULL_PARAM_DEREF_NONVIRTUAL,NP_BOOLEAN_RETURN_NULL,CT_CONSTRUCTOR_THROW,NP_GUARANTEED_DEREF_ON_EXCEPTION_PATH,NP_NULL_PARAM_DEREF_ALL_TARGETS_DANGEROUS,EI_EXPOSE_REP,OBL_UNSATISFIED_OBLIGATION,EI_EXPOSE_REP2,DMI_EMPTY_DB_PASSWORD,EQ_DOESNT_OVERRIDE_EQUALS,NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE,ST_WRITE_TO_STATIC_FROM_INSTANCE_METHOD,DM_DEFAULT_ENCODING" />
+  </Match>
+
+</FindBugsFilter>
