--- laokou-common/laokou-common-grpc/src/main/java/org/laokou/common/grpc/config/GrpcClientConfig.java
@@ -1,65 +1,0 @@
-/*
- * Copyright (c) 2022-2026 KCloud-Platform-IoT Author or Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.laokou.common.grpc.config;
-
-import io.grpc.netty.NettyChannelBuilder;
-import org.jspecify.annotations.NonNull;
-import org.laokou.common.grpc.annotation.GrpcClientBeanPostProcessor;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.cloud.client.discovery.DiscoveryClient;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.grpc.client.ClientInterceptorsConfigurer;
-import org.springframework.grpc.client.GrpcChannelBuilderCustomizer;
-import org.springframework.grpc.client.GrpcClientFactory;
-import org.springframework.grpc.client.ImportGrpcClients;
-
-import java.util.List;
-
-/**
- * @author laokou
- */
-@ImportGrpcClients
-@Configuration(proxyBeanMethods = false)
-@ConditionalOnClass(DiscoveryClient.class)
-public class GrpcClientConfig {
-
-	@Bean
-	DiscoveryNameResolverProvider discoveryNameResolverProvider(DiscoveryClient discoveryClient) {
-		return new DiscoveryNameResolverProvider(discoveryClient);
-	}
-
-	@Bean
-	DiscoveryNameResolverRegister discoveryNameResolverRegister(
-			DiscoveryNameResolverProvider discoveryNameResolverProvider) {
-		return new DiscoveryNameResolverRegister(discoveryNameResolverProvider);
-	}
-
-	@Bean
-	GrpcClientBeanPostProcessor grpcClientBeanPostProcessor(GrpcClientFactory grpcClientFactory) {
-		return new GrpcClientBeanPostProcessor(grpcClientFactory);
-	}
-
-	@Bean
-	DiscoveryGrpcChannelFactory discoveryGrpcChannelFactory(
-			List<GrpcChannelBuilderCustomizer<@NonNull NettyChannelBuilder>> globalCustomizers,
-			ClientInterceptorsConfigurer interceptorsConfigurer) {
-		return new DiscoveryGrpcChannelFactory(globalCustomizers, interceptorsConfigurer);
-	}
-
-}
