--- laokou-common/laokou-common-grpc/src/main/java/org/laokou/common/grpc/exception/ServiceNotFoundException.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2022-2026 KCloud-Platform-IoT Author or Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.laokou.common.grpc.exception;
-
-import org.laokou.common.i18n.common.exception.BizException;
-
-/**
- * @author laokou
- */
-public class ServiceNotFoundException extends BizException {
-
-	public ServiceNotFoundException(String code, String msg, Throwable throwable) {
-		super(code, msg, throwable);
-	}
-
-}
