--- laokou-common/laokou-common-grpc/src/test/java/org/laokou/common/grpc/config/DiscoveryGrpcChannelFactoryTest.java
@@ -1,59 +1,0 @@
-/*
- * Copyright (c) 2022-2026 KCloud-Platform-IoT Author or Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.laokou.common.grpc.config;
-
-import io.grpc.ChannelCredentials;
-import io.grpc.InsecureChannelCredentials;
-import io.grpc.netty.NettyChannelBuilder;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-import org.springframework.grpc.client.ClientInterceptorsConfigurer;
-
-import java.util.Collections;
-
-/**
- * @author laokou
- */
-class DiscoveryGrpcChannelFactoryTest {
-
-	private DiscoveryGrpcChannelFactory factory;
-
-	@BeforeEach
-	void setUp() {
-		ClientInterceptorsConfigurer configurer = Mockito.mock(ClientInterceptorsConfigurer.class);
-		factory = new DiscoveryGrpcChannelFactory(Collections.emptyList(), configurer);
-	}
-
-	@Test
-	void testSupports() {
-		Assertions.assertThat(factory.supports("discovery:laokou-auth")).isTrue();
-		Assertions.assertThat(factory.supports("static:localhost")).isFalse();
-		Assertions.assertThat(factory.supports("dns:localhost")).isFalse();
-	}
-
-	@Test
-	void testNewChannelBuilder() {
-		String target = "laokou-auth";
-		ChannelCredentials credentials = InsecureChannelCredentials.create();
-		NettyChannelBuilder builder = factory.newChannelBuilder(target, credentials);
-		Assertions.assertThat(builder).isNotNull();
-	}
-
-}
