+++ laokou-common/laokou-common-log/pom.xml
@@ -28,6 +28,11 @@
       <artifactId>laokou-snowflake-id-proto</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.laokou</groupId>
+      <artifactId>laokou-common-grpc-client</artifactId>
+      <version>${project.version}</version>
+    </dependency>
   </dependencies>
 
 </project>
