+++ laokou-service/laokou-auth/laokou-auth-infrastructure/pom.xml
@@ -147,6 +147,11 @@
       <artifactId>laokou-snowflake-id-proto</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.laokou</groupId>
+      <artifactId>laokou-common-grpc-client</artifactId>
+      <version>${project.version}</version>
+    </dependency>
   </dependencies>
 
 </project>
